Move extension to libadcc
