--- LICENSE_adccore
@@ -1,140 +1,0 @@
-######################
-#--  Introduction  --#
-######################
-adcc and the adcc source code is released under the terms of the GNU General Public
- License Version 3 (GPLv3). This terms of release, however, do not apply to the
-adccore binary file `libadccore.so` (or a simlilarly named dylib file for MacOS),
-contained in the `adcc/lib` directory of the python code tarball or the python
-package directory.
-
-This measure was taken, since adccore integrates dependencies, where the
-licence terms only permit a public release in the form of a precompiled binary.
-The precise terms, which apply to the binary library blob adcc/lib/libadccore.so
-or adcc/lib/libadccore.dylib can be found in the next sections of this file.
-A list of components compiled into libadccore is given in the last section.
-
-
-#################################################################
-#--  adccore software licence agreement and limited warranty  --#
-#################################################################
-This legal document is an agreement between you, THE LICENSEE, and the
-adcc developers, THE LICENSOR. By executing the adccore library by
-itself or dynamically loaded from another software, you agree to be bound by the
-terms of this agreement, which includes the Software License, Limited Warranty
-and Acknowledgment. If you do not agree to the terms presented in this
-agreement, you may not execute adccore. Limited by the terms defined
-in the Software license shown below, the adcc developers grant THE LICENSEE a
-simple, non-exclusive, non-transferable and non-sublicensable right to use
-the software. Use in accordance with the Agreement includes installation of
-adccore and loading, display and running of the installed library.
-LICENSEE may duplicate adccore insofar as such duplication is required to use adccore.
-
-
-##########################
-#--  Software license  --#
-##########################
-1. **Grant of a License.** Given your agreement to abide by the terms and
-conditions of this License and the Limited Warranty, the adcc developers,
-THE LICENSOR, grants you, THE LICENSEE, a non-exclusive right to use this copy
-of adccore so long as you comply with the terms of this License.
-
-2. **Ownership of Software.** The adcc developers retain title and ownership of
-the original copy of the Software and all subsequent copies of the
-adccore, regardless of the form or media in or on which the original
-and other copies may exist.
-
-3. **Use Restrictions.** As the Licensee, you may physically transfer
-adccore from one computer to another and distribute copies of
-adccore. You may, however, not modify, adapt,  translate,
-reverse engineer, disassemble, decompile, or create derivative works based
-on adccore.
-
-4. **Limitations on Reverse Engineering, Decompilation, and Disassembly.**
-You may not reverse engineer, decompile, or disassemble adccore,
-except and only to the extent that such activity is expressly permitted by
-applicable law notwithstanding this limitation.
-
-5. **Limitations on Separation of Components.** adccore is licensed
-as a single product. Its component parts may not be separated for use on more
-than one software at once.
-
-6 **Limitations on Rental or Lease. You may not rent or lease adccore.
-
-7. **Limitations on Software Transfer.** You may not transfer your rights under
- this License.
-
-8. **Termination.** Without prejudice to any other rights the adcc developers
-may terminate this License if you fail to comply with the terms and conditions
-of this License. In such event, you must destroy all copies of adccore
-and all of its component parts.
-
-9. No relaxation or indulgence which any party may grant to any other party
-shall constitute a waiver of the rights of that party and shall not preclude
-that party from exercising any rights which may have arisen in the past or which
-might arise in future.
-
-8. Miscellaneous. This License shall be governed and construed inaccordance with
-the laws of the Federal Republic of Germany. Should any part of the Agreement be
-held invalid, the remainder of the Agreement will still be in effect.
-
-
-#####################################################
-#--  Limited Warranty and Disclaimer of Warranty  --#
-#####################################################
-adccore is provided "as is" without warranty of any kind. Further
-the adcc developers do not warrant, guarantee or make any representations
-regarding the use or the results of use, of adccore in terms of
-correctness, reliability, accuracy or fitness for purpose. The entire risk as to
-the results and performance of the software is assumed by you. Should the
-software be defective, you and not the adcc developers assume the entire cost of
-all necessary repair, servicing and correction.
-
-The above is the only warranty of any kind, either express or implied, including
-but not limited to the warranties of merchantability and fitness for purpose,
-that the adcc develpers make regarding this product. No oral or written
-information or advise given by the adcc developers shall create a warranty or in
-any way increase the scope of this warranty.
-
-Further the adcc developers shall not be liable for any direct, indirect,
-consequential or incidental damages ( including damages for loss of business
-profits, business interruption, loss of business information, and the like.)
-arising out of the use or inability to use this product.
-
-This does not affect your statutory rights.
-
-This Limited Warranty shall be governed and construed in accordance with the
-laws of the Federal Republic of Germany.
-
-
-########################
-#--  Acknowledgment  --#
-########################
-You acknowledge that you have read this License and Limited Warranty, understand
-them, and agree to be bound by their terms and conditions.
-
-
-####################################################
-#--  Overview of adccore third-party components  --#
-####################################################
-This section lists the third-party components integrated into adccore in binary
-form, provides the copyright holders and their terms of use.
-
-#
-# libadc
-#
-- **Description:** Library of ADC equations in libtensor syntax.
-- **Copyright:**   Copyright (c) 2015 Michael Wormit, Adrian L. Dempwolff,
-                   Philipp H. P. Harbach, Caroline Krauter, Dirk Rehn,
-                   Matthias Schneider and Jan Wenzel
-- **Publication:** http://doi.org/10.1080/00268976.2013.859313
-
-
-#
-# libtensor
-#
-- **Description:** C++ library for tensor computations.
-- **Copyright:**   Copyright (c) 2013 Evgeny Epifanovsky, Michael Wormit,
-                   Dmitry Zuev, Sam Manzer, Ilya Kaliman
-- **License:**     Boost Software License 1.0
-- **Website:**     https://github.com/epifanovsky/libtensor
-- **Publication:** http://dx.doi.org/10.1002/jcc.23377
