--- adcc/lib/.gitignore
@@ -1,5 +1,0 @@
-/*.so
-/*.so.*
-/*.dylib
-/*.a
-/libadccore_thirdparty
