--- extension/adccore/.gitignore
@@ -1,3 +1,0 @@
-/*
-!/.gitignore
-!/README.md
