--- extension/adccore/README.md
@@ -1,1 +1,0 @@
-Untar or install adccore into this directory
