--- extension/adccore/lib
@@ -1,1 +1,0 @@
-../../adcc/lib
