+++ libadcc/convert_timer.cc
