+++ libadcc/convert_timer.hh
