+++ libadcc/export_AdcMemory.cc
