+++ libadcc/export_MoIndexTranslation.cc
