+++ libadcc/export_MoSpaces.cc
