+++ libadcc/export_ReferenceState.cc
