+++ libadcc/export_Symmetry.cc
