+++ libadcc/export_amplitude_vector_enforce_spin_kind.cc
