+++ libadcc/export_guesses.cc
