+++ libadcc/export_threading.cc
