+++ libadcc/hartree_fock_solution_hack.hh
