+++ libadcc/util.cc
