+++ libadcc/util.hh
